Day 1 Hi Stoyan Support Long Long implemented

+++ solutions/easy/main.cpp
@@ -4,13 +4,13 @@ using namespace std;
 
 int main()
 {
-    int a, b;
+    int long long a, b;
     cout << "Enter first number: " << endl;
     cin >> a;
     cout << "Enter second number:" << endl;
     cin >> b;
 
-    int sum = a + b;
+    int long long sum = a + b;
     cout << "the sum is: " << sum << endl;
     return 0;
 }
